Javascript music playlist web-app
This Web application is a Javascript project, a music playlist application that plays your favorite songs, next song, or the previous song. It will not end playing the song and once the playlist reaches the end and you hit next it will jump back to the first song of the playlist again.